documentation and consistency
