+++ package.json
@@ -23,7 +23,8 @@
         "crc": "^3.8.0",
         "fs-extra": "^8.0.1",
         "merge": "^1.2.1",
-        "qs": "^6.7.0"
+        "qs": "^6.7.0",
+        "express": "^4.16.4"
     },
     "devDependencies": {
         "@babel/core": "^7.4.4",
@@ -34,7 +35,6 @@
         "eslint-config-prettier": "^4.2.0",
         "eslint-plugin-import": "^2.17.2",
         "eslint-plugin-jest": "^22.6.4",
-        "express": "^4.16.4",
         "jest": "^24.7.1"
     },
     "main": "./src/index.js"
