+++ test/config.test.js
@@ -1,5 +1,5 @@
 const H5PEditorConfig = require('../src/config');
-const InMemoryStorage = require('../src/in-memory-storage');
+const InMemoryStorage = require('./mockups/in-memory-storage');
 
 describe('loading configuration data from storage', () => {
     it('doesn\'t overwrite defaults if storage is empty', async () => {
